pyproject.toml
[project]
name = "manim-gpt"
version = "0.1.0"
description = "AI-powered Manim video generation API with multi-format support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.2",
    "litellm>=1.80.0",
    "pydantic>=2.12.4",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0",
    "manim>=0.18.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "gradio>=4.0.0",
    "httpx>=0.24.0",
]
